Move gluon-apparatus image into this codebase
It doesn't need to exist outside of this project, so the OCI image
recipe for building the job implementation-y bits can now be found under
docker/gluon-apparatus.  The API machinery has been updated to point to
those things, unless someone overrides the image name/tag with
environment variables.

Note: we also now default to only pulling the image if it is not already
present, since the apparatus image has calmed down considerably since
the start of this thing last week.  That can also be overridden by
environment variables.